+++ .idea/workspace.xml
@@ -3,6 +3,19 @@
   <component name="ChangeListManager">
     <list default="true" id="a190d58c-ffb4-42f1-bf9d-1d35f65a9c00" name="Default Changelist" comment="">
       <change beforePath="$PROJECT_DIR$/../SY-GitLab-api/shopping-api-app" beforeDir="false" afterPath="$PROJECT_DIR$/../SY-GitLab-api/shopping-api-app" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/Circle.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/Circle.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/Loophole/Loophole .py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/Loophole/Loophole .py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/bug/bug.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/bug/bug.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/code/code_number.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/code/code_number.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/debt_rate.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/debt_rate.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/debt_time.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/debt_time.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/hot/hotspotsecure.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/hot/hotspotsecure.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/odor/odor.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/odor/odor.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/repeat/repeat_block.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/repeat/repeat_block.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/repeat/repeat_rate.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/repeat/repeat_rate.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/SonarQube/review_rate.py" beforeDir="false" afterPath="$PROJECT_DIR$/SonarQube/review_rate.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/other/test.java" beforeDir="false" afterPath="$PROJECT_DIR$/other/test.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/../robot/wujian/robot/.idea/master.iml" beforeDir="false" afterPath="$PROJECT_DIR$/../robot/wujian/robot/.idea/master.iml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/../robot/wujian/robot/.idea/modules.xml" beforeDir="false" afterPath="$PROJECT_DIR$/../robot/wujian/robot/.idea/modules.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/../robot/wujian/robot/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/../robot/wujian/robot/.idea/workspace.xml" afterDir="false" />
@@ -36,8 +49,11 @@
   </component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
-      <recent name="$PROJECT_DIR$/../Practice/ONESapi/update" />
-      <recent name="$PROJECT_DIR$" />
+      <recent name="$PROJECT_DIR$/SonarQube" />
+      <recent name="$PROJECT_DIR$/SonarQube/repeat" />
+      <recent name="$PROJECT_DIR$/SonarQube/leak" />
+      <recent name="$PROJECT_DIR$/SonarQube/hot" />
+      <recent name="$PROJECT_DIR$/SonarQube/flavor" />
     </key>
     <key name="CopyFile.RECENT_KEYS">
       <recent name="$PROJECT_DIR$/../robot/desk-issue/wecha-reminder/ones-tools/wechat-reminder" />
@@ -56,7 +72,28 @@
       </list>
     </option>
   </component>
-  <component name="RunManager" selected="Python.Hello">
+  <component name="RunManager" selected="Python.Circle">
+    <configuration name="Circle" type="PythonConfigurationType" factoryName="Python" temporary="true">
+      <module name="TestDveOps" />
+      <option name="INTERPRETER_OPTIONS" value="" />
+      <option name="PARENT_ENVS" value="true" />
+      <envs>
+        <env name="PYTHONUNBUFFERED" value="1" />
+      </envs>
+      <option name="SDK_HOME" value="" />
+      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/SonarQube" />
+      <option name="IS_MODULE_SDK" value="true" />
+      <option name="ADD_CONTENT_ROOTS" value="true" />
+      <option name="ADD_SOURCE_ROOTS" value="true" />
+      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/Circle.py" />
+      <option name="PARAMETERS" value="" />
+      <option name="SHOW_COMMAND_LINE" value="false" />
+      <option name="EMULATE_TERMINAL" value="false" />
+      <option name="MODULE_MODE" value="false" />
+      <option name="REDIRECT_INPUT" value="false" />
+      <option name="INPUT_FILE" value="" />
+      <method v="2" />
+    </configuration>
     <configuration name="Hello" type="PythonConfigurationType" factoryName="Python" temporary="true">
       <module name="TestDveOps" />
       <option name="INTERPRETER_OPTIONS" value="" />
@@ -69,7 +106,7 @@
       <option name="IS_MODULE_SDK" value="true" />
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
-      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/Count_date.py" />
+      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/SonarQube/bug/bug.py" />
       <option name="PARAMETERS" value="" />
       <option name="SHOW_COMMAND_LINE" value="false" />
       <option name="EMULATE_TERMINAL" value="false" />
@@ -91,6 +128,27 @@
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
       <option name="SCRIPT_NAME" value="" />
+      <option name="PARAMETERS" value="" />
+      <option name="SHOW_COMMAND_LINE" value="false" />
+      <option name="EMULATE_TERMINAL" value="false" />
+      <option name="MODULE_MODE" value="false" />
+      <option name="REDIRECT_INPUT" value="false" />
+      <option name="INPUT_FILE" value="" />
+      <method v="2" />
+    </configuration>
+    <configuration name="repeat_block" type="PythonConfigurationType" factoryName="Python" temporary="true">
+      <module name="TestDveOps" />
+      <option name="INTERPRETER_OPTIONS" value="" />
+      <option name="PARENT_ENVS" value="true" />
+      <envs>
+        <env name="PYTHONUNBUFFERED" value="1" />
+      </envs>
+      <option name="SDK_HOME" value="" />
+      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/SonarQube/repeat" />
+      <option name="IS_MODULE_SDK" value="true" />
+      <option name="ADD_CONTENT_ROOTS" value="true" />
+      <option name="ADD_SOURCE_ROOTS" value="true" />
+      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/SonarQube/repeat/repeat_block.py" />
       <option name="PARAMETERS" value="" />
       <option name="SHOW_COMMAND_LINE" value="false" />
       <option name="EMULATE_TERMINAL" value="false" />
@@ -186,6 +244,8 @@
     </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="Python.Circle" />
+        <item itemvalue="Python.repeat_block" />
         <item itemvalue="Python.Hello" />
         <item itemvalue="Python.weely" />
         <item itemvalue="Python.test_api" />
