+++ SonarQube/code/code_number.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-# 
 #-------------------------------------------------------------------------------
-# Name:         code_number
-# Description:  
-# Author:       wujian
-# Date:         2020/6/18
-#-------------------------------------------------------------------------------
+
+code = "代码行数"
+
+for code_str in code:
+    print(code_str)
