+++ SonarQube/debt_rate.py
@@ -1,7 +1,4 @@
 # -*- coding: utf-8 -*-# 
 #-------------------------------------------------------------------------------
-# Name:         debt_ratio.py
-# Description:  
-# Author:       wujian
-# Date:         2020/6/18
-#-------------------------------------------------------------------------------
+debt_rate = "债务率"
+
