+++ SonarQube/debt_time.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-#
+#-------------------------------------------------------------------------------
+debt_rate = "债务时间"
+
