+++ SonarQube/hot/hotspotsecure.py
@@ -1,3 +1,14 @@
 # -*- coding: utf-8 -*-# 
 #-------------------------------------------------------------------------------
 
+hotspotsecure = "安全热点"
+DATABASES = {
+    'postgresql_db': {
+        'ENGINE': 'django.db.backends.postgresql',
+        'NAME': 'quickdb',
+        'USER': 'sonarsource',
+        'PASSWORD': '123456',
+        'HOST': 'localhost',
+        'PORT': '5432'
+    }
+}
