+++ SonarQube/repeat/repeat_rate.py
@@ -1,7 +1,33 @@
 # -*- coding: utf-8 -*-# 
 #-------------------------------------------------------------------------------
-# Name:         repetition_rate
-# Description:  
-# Author:       wujian
-# Date:         2020/6/18
-#-------------------------------------------------------------------------------
+
+
+repeat_rate = "代码：{}".format("重复率")
+
+
+def count_number(start):
+    add_number = 0
+    add_number += int(start)
+    print(count_number)
+    return count_number
+
+def count_str(start):
+    add_number = "str"
+    add_number += start
+    count_str = len(add_number)
+    print(count_str)
+    return count_str
+
+def count_list(start):
+    add_number = ["list01","list02"]
+    add_number.append(start)
+    count_list = len(add_number)
+    return count_list
+
+if __name__ == '__main__':
+    count_number(2)
+    count_str("str1")
+    count_list("ces")
+
+
+
