+++ SonarQube/review_rate.py
@@ -1,7 +1,4 @@
 # -*- coding: utf-8 -*-# 
 #-------------------------------------------------------------------------------
-# Name:         code_review
-# Description:  
-# Author:       wujian
-# Date:         2020/6/18
-#-------------------------------------------------------------------------------
+review_rate = "复审率"
+
